+++ experiments/yolo_experiment.py
@@ -4,7 +4,7 @@ from typing import Dict, Any
 import torch
 
 # Add the project root to the Python path
-project_root = Path(__file__).resolve().parents[3]
+project_root = Path(__file__).resolve().parent.parent
 if str(project_root) not in sys.path:
     sys.path.append(str(project_root))
